Show promo on ticket purchase and applies promo to stripe payment.

+++ OpenEvent.Web/Models/Promo/CreatePromoBody.cs
@@ -7,7 +7,7 @@ namespace OpenEvent.Web.Models.Promo
         public DateTime Start { get; set; }
         public DateTime End { get; set; }
         public bool Active { get; set; }
-        public double Discount { get; set; }
+        public int Discount { get; set; }
         public Guid EventId { get; set; }
     }
 }
